// src/html_editor/types/serializedTheme.ts
export type SerializedThemeType = {
    id: string
    project_id: string
    name: string
    // palette
    primary_main: string
    primary_light: string
    primary_dark: string
    primary_contrasttext: string
    secondary_main: string
    secondary_light: string
    secondary_dark: string
    secondary_contrasttext: string
    error_main: string
    error_light: string
    error_dark: string
    error_contrasttext: string
    warning_main: string
    warning_light: string
    warning_dark: string
    warning_contrasttext: string
    info_main: string
    info_light: string
    info_dark: string
    info_contrasttext: string
    success_main: string
    success_light: string
    success_dark: string
    success_contrasttext: string
    text_primary: string
    text_secondary: string
    text_disabled: string
    background_default: string
    background_paper: string
    action_active: string
    action_hover: string
    action_selected: string
    action_disabled: string
    action_disabled_background: string
    action_focus: string
    mode: string
  }